# static/specs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rainbow Authentication Authority API",
    "description": "Rainbow Authentication Module - Rainbow Authority API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1000",
      "description": "Testing server"
    }
  ],
...
  "paths": {
...
    "/api/v1/continue/{id}": {
      "post": {
        "tags": [
          "Minions"
        ],
        "summary": "Continue GNAP request",
        "description": "Continues a GNAP access request using the interaction reference and Authorization token.",
        "operationId": "continue_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Request identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "interact_ref": {
                    "type": "string"
                  }
                },
                "required": [
                  "interact_ref"
                ]
              }
            }
          }
        },
        "headers": {
          "Authorization": {
            "description": "GNAP access token, format: 'GNAP <token>'",
            "schema": {
              "type": "string"
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "manage": {
                      "type": "object"
                    },
                    "access": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "expires_in": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "key": {
                      "type": "object"
                    },
                    "flags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "value",
                    "access"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: missing or invalid Authorization token."
          },
          "404": {
            "description": "Request with given id not found."
          },
          "422": {
            "description": "Unprocessable entity: interact_ref or token does not match expected value."
          },
          "500": {
            "description": "Unexpected internal server error or database error."
          }
        }
      }
    },
    "/api/v1/pd/{state}": {
      "get": {
        "tags": [
          "OIDC4VP"
        ],
        "summary": "Retrieve VP Definition",
        "description": "Generates and retrieves the Verifiable Presentation definition for the given state identifier.",
        "operationId": "pd",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "description": "State identifier for the VP request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verifiable Presentation definition successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "State not found: no VP Definition exists for the given identifier."
          },
          "500": {
            "description": "Unexpected internal server error or database error."
          }
        }
      }
    },
    "/api/v1/verify/{state}": {
      "post": {
        "tags": [
          "OIDC4VP"
        ],
        "summary": "Verify VP Submission",
        "description": "Verifies the submitted Verifiable Presentation for a given state identifier.",
        "operationId": "verify",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "description": "State identifier for the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "vp_token": {
                    "type": "string"
                  },
                  "presentation_submission": {
                    "type": "string"
                  }
                },
                "required": [
                  "vp_token",
                  "presentation_submission"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification completed, returns URI if available.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request: malformed or missing fields."
          },
          "404": {
            "description": "State not found: no verification request exists for the given identifier."
          },
          "422": {
            "description": "Unprocessable entity: validation failed."
          },
          "500": {
            "description": "Unexpected internal server error or database error."
          }
        }
      }
    },
...
  }
}